requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
soupsieve==2.6
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
